修复按钮位置计算并切换 WebSocket 目标地址
修改详情：
- 在 `CGoofishDlg::OnInitDialog()` 方法中，添加显式的 `(int)` 类型转换，确保按钮位置和间距计算结果为整数类型。
- 在 `CGoofishDlg::OnBtnStart()` 方法中，将 WebSocket 连接地址从测试服务器切换为正式服务器 `wss://wss-goofish.dingtalk.com`。

升级工具集版本并优化字体高度计算

在 `Goofish.vcxproj` 文件中，将所有配置的 `<PlatformToolset>` 属性从 `v143` 升级为 `v145`，以更新项目的编译工具集版本。

在 `FontModel.cpp` 文件中，调整了 `fontHeight` 的计算方式，增加了显式的类型转换 `(int)`，以确保类型安全性。

+++ Goofish/Model/FontModel.cpp
@@ -3,7 +3,7 @@
 
 void FontModel::OnInit()
 {
-	int fontHeight = 13 * Utils::GetDpi(); // 96 是标准 DPI
+	int fontHeight = (int)(13 * Utils::GetDpi()); // 96 是标准 DPI
 
 	m_font.CreateFont(
 		fontHeight,                // 字体高度
